+++ src/entities/Bullet.cpp
@@ -3,13 +3,13 @@
 #include "components/Line.hpp"
 #include "components/game/BulletInfo.hpp"
 
-void makeBullet(entt::registry &reg, const Vec2d &from, const Vec2d &to, double speed)
+void makeBullet(entt::registry &reg, const Vec2f &from, const Vec2f &to, float speed)
 {
     const entt::entity e = reg.create();
 
     reg.emplace<BulletInfo>(e, from, to, speed);
 
-    const Vec2d lineEnd = from + normalize(Vec2d{to.x - from.x, to.y - from.y}) * 30;
+    const Vec2f lineEnd = from + normalize(Vec2f{to.x - from.x, to.y - from.y}) * 30;
 
     reg.emplace<Line>(e, from, lineEnd, false);
 }
